+++ MNE/fiff/fiff_named_matrix.cpp
@@ -64,12 +64,12 @@ FiffNamedMatrix::FiffNamedMatrix()
 
 //*************************************************************************************************************
 
-FiffNamedMatrix::FiffNamedMatrix(fiff_int_t p_nrow, fiff_int_t p_ncol, QStringList& p_row_names, QStringList& p_col_names, MatrixXf* p_data)
+FiffNamedMatrix::FiffNamedMatrix(fiff_int_t p_nrow, fiff_int_t p_ncol, QStringList& p_row_names, QStringList& p_col_names, MatrixXd* p_data)
 : nrow(p_nrow)
 , ncol(p_ncol)
 , row_names(p_row_names)
 , col_names(p_col_names)
-, data(p_data ? new MatrixXf(*p_data) : NULL)
+, data(p_data ? new MatrixXd(*p_data) : NULL)
 {
 }
 
@@ -81,7 +81,7 @@ FiffNamedMatrix::FiffNamedMatrix(const FiffNamedMatrix* p_pFiffNamedMatrix)
 , ncol(p_pFiffNamedMatrix->ncol)
 , row_names(p_pFiffNamedMatrix->row_names)
 , col_names(p_pFiffNamedMatrix->col_names)
-, data(p_pFiffNamedMatrix->data ? new MatrixXf(*p_pFiffNamedMatrix->data) : NULL)
+, data(p_pFiffNamedMatrix->data ? new MatrixXd(*p_pFiffNamedMatrix->data) : NULL)
 {
 }
